Combine timestamp.rs and delay.rs into time.rs in types
