+++ packages/types/src/lib.rs
@@ -1,20 +1,18 @@
 mod aggregator_types;
 mod bytes;
 mod chain_config;
-mod duration;
 mod http;
 mod id;
 mod packet;
 mod service;
 mod solidity_types;
-mod timestamp;
+mod time;
 pub use aggregator_types::*;
 pub use bytes::*;
 pub use chain_config::*;
-pub use duration::*;
 pub use http::*;
 pub use id::*;
 pub use packet::*;
 pub use service::*;
 pub use solidity_types::*;
-pub use timestamp::*;
+pub use time::*;
